publish jdcloud-sdk-go version 1.1.2

+++ services/clouddnsservice/apis/GetDomainQueryCount.go
@@ -153,6 +153,6 @@ type GetDomainQueryCountResponse struct {
 }
 
 type GetDomainQueryCountResult struct {
-    Time []int `json:"time"`
-    Traffic []int `json:"traffic"`
+    Time []int64 `json:"time"`
+    Traffic []int64 `json:"traffic"`
 }
